+++ apps/nextjs/src/components/AppComponents/Chat/Quiz/QuizQuestionsListItem/index.tsx
@@ -1,2 +1,0 @@
-export { default } from "./QuizQuestionsListItem";
-export type { QuizQuestionsListItemProps as QuestionListItemProps } from "./QuizQuestionsListItem";